dock: removed own __NOTIFYICONDATAA/W structures icq: removed last parts of invisible check rest: cleanups
git-svn-id: file:///.data/svn_convert/svn@6803 f7a67266-ba00-4dfb-90a2-3eec4ba07b7d

+++ branches/branch-QString/plugins/icq/pastinfo.cpp
@@ -124,7 +124,7 @@ void PastInfo::fill()
     while (str.length()){
         QString info = getToken(str, ';', false);
         QString n = getToken(info, ',');
-        unsigned short category = (unsigned short)n.toLong();
+        unsigned short category = n.toUShort();
         switch (i){
         case 0:
             edtBg1->setText(info);
@@ -159,7 +159,7 @@ void PastInfo::fill()
     while (str.length()){
         QString info = getToken(str, ';', false);
         QString n = getToken(info, ',');
-        unsigned short category = (unsigned short)n.toLong();
+        unsigned short category = n.toUShort();
         switch (i){
         case 0:
             edtAf1->setText(info);
